Added turn based logic and inputs added

+++ Base/Source/Main/Game/Objects/Characters/MeleeCharacter.cpp
@@ -12,8 +12,8 @@ MeleeCharacter::~MeleeCharacter()
 
 void MeleeCharacter::SetCharacter(UnitType* Type, UnitRace* Race)
 {
-	HealthPoints = Type->GetHealth() * Race->GetHealthModifier();
-	MaxHealthPoints = Type->GetMaxHealth() * Race->GetHealthModifier();
+	HealthPoints = (int)(Type->GetHealth() * Race->GetHealthModifier());
+	MaxHealthPoints = (int)(Type->GetMaxHealth() * Race->GetHealthModifier());
 	WalkSpeed = Type->GetWalkspeed();
 	DetectionRadius = Type->GetRange();
 	RaceType = Race->GetRace();
@@ -24,7 +24,7 @@ void MeleeCharacter::SetCharacter(UnitType* Type, UnitRace* Race)
 	WT = Type->PossibleWeapon[RandomChoice];
 
 	MeleeStateManager* MS = dynamic_cast<MeleeStateManager*>(InternalStateManager);
-	MS->MWeapon = new MeleeWeapon(WT->GetDamage() * Race->GetMeleeDamageModifier(), WT->GetRate(), 2);
+	MS->MWeapon = new MeleeWeapon((int)(WT->GetDamage() * Race->GetMeleeDamageModifier()), WT->GetRate(), 2);
 
 }
